Rename holidy api package
