+++ src/main/java/org/synyx/urlaubsverwaltung/holiday/PublicHolidayConfiguration.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.holiday.api;
+package org.synyx.urlaubsverwaltung.holiday;
 
 import de.jollyday.HolidayManager;
 import de.jollyday.ManagerParameter;
