+++ src/main/java/org/synyx/urlaubsverwaltung/holiday/PublicHolidaysDto.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.holiday.api;
+package org.synyx.urlaubsverwaltung.holiday;
 
 import java.util.ArrayList;
 import java.util.List;
